# db: optimize levelIter for repeated seeks with narrow and monotonic bounds