--- a/testdata/level_iter
+++ b/testdata/level_iter
@@ -489,7 +489,7 @@ set-bounds lower=cc
 seek-lt d
 ----
 d#4,1:4
-c#3,15:
+.
 
 iter
 seek-ge c
@@ -497,4 +497,4 @@ set-bounds upper=cc
 seek-ge d
 ----
 c#3,1:3
-d#4,15:
+.